* Rename mn_compositor_node_group.py to compositor.py
